+++ src/lib/convenience/index.ts
@@ -1,1 +1,0 @@
-export { default as WeekSchedule } from './WeekSchedule.svelte';